feat: Install card-swiper && Add main banner, taps section
Card-swiper is the most popular package for swipe gesture. It was so easy to use.
Taps sections consist with ToggleButtons cause apply for active style. Untill now
it is not going to work. Maybe there's another way to update list with setState method.
So i'm working on and on and on...