feat: Add model class for fetched data
In dart, fetched data should be converted to intance of model class. Each of model class will have
'fromJson' and 'toJson' method that parsing json to map and map to json. Converted instances will be
collected to Store model class.

Store model class is called by AppController which mediates eacth of widgets and models. It has
responsibilty about model, later it is going to be moderator all about thing.